\docType{package}
\name{calpr-package}
\alias{calpr-package}
\alias{calpr}

\title{Calendar Periods}

\description{
A small package for working with calendar periods (e.g., year-months, year-quarters) using base R dates.
}

\details{
This package has essentially two functions for working with base R dates: \code{\link[calpr:calendar_period]{calendar_period()}} turns dates into calendar periods (e.g., year-months), and \code{\link[calpr:complete_periods]{complete_periods()}} checks if dates form complete calendar periods.
}

\seealso{
\url{https://github.com/marberts/gpindex}
}

\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com}
}
